+++ packages/geospatial/__init__.py
@@ -1,5 +1,5 @@
 from .atlas import AtlasResult, AtlasService
-from .live_adapters import CensusGeocoderAdapter
+from .live_adapters import CensusGeocoderAdapter, USGSWatershedAdapter
 from .privacy import Coordinate, PrivacyReducedCoordinate, reduce_coordinate_precision
 from .tools import CensusGeographyTool, admin_resolution_from_tool_result
 
@@ -8,6 +8,7 @@ __all__ = [
     "AtlasService",
     "CensusGeocoderAdapter",
     "CensusGeographyTool",
+    "USGSWatershedAdapter",
     "Coordinate",
     "PrivacyReducedCoordinate",
     "admin_resolution_from_tool_result",